requirements.txt
msgpack_python==0.5.6
numpy==2.1.1
opencv_python==4.9.0.80
Pillow==10.4.0
psutil==5.9.8
pygame==2.5.2
